refactor: decouple project index metrics
